pyproject.toml
[project]
name = "dteam-django-practical-test"
version = "0.1.0"
description = ""
authors = [
    {name = "Your Name",email = "you@example.com"}
]
readme = "README.md"
requires-python = ">=3.12, <4"
dependencies = [
    "django (>=5.2.1,<6.0.0)",
    "xhtml2pdf (>=0.2.17,<0.3.0)",
    "djangorestframework (>=3.16.0,<4.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
